refactor:PuppyDogAnimationの状態管理分離、型定義
